Extract read-error recovery, fix the mount wait, drop dead code
rip.service.js was 1,388 lines across four jobs. The read-error recovery
workflow - imaging a damaged disc, re-ripping from the image, image locking,
free-space and retention handling - is now ReadErrorRecovery in its own module
(898 + 584 lines). It needs exactly two things from the ripper, so both are
injected: a cancellation seam and a callback for handing recovered files to the
encode queue. Its tests target it directly instead of reaching through
RipService, which is what they were really testing all along.

That also removed a duplicated block: queueing MKVs for HandBrake existed twice
with slightly different guards. One queueForEncoding method now serves the
normal rip path and recovery.

waitForDriveMount had two problems. It counted poll attempts rather than
elapsed time, and each attempt costs two makemkvcon invocations - so on a slow
drive a "10 second" wait could run for minutes. It is wall-clock now. It also
tracked "discs new since we started", with a baseline that a failed first poll
left at zero, making every disc look new; since the caller merges by drive
number anyway, it simply reports everything it can see. That deletes the
baseline bookkeeping and one makemkvcon call per wait.

NativeOpticalDrive.isNativeAvailable threw on the first call when the addon was
missing and returned false on every call after, so the same failure produced
two different errors depending on ordering. It answers false consistently now.
Also dropped: its unused ejectAllDrives/loadAllDrives (OpticalDriveUtil does its
own iteration), four `try { ... } catch (e) { throw e }` wrappers, and
getDefaultHandBrakeConfig/mergeHandBrakeConfig, a second source of truth for
HandBrake defaults that nothing called.

Verified the app still boots and serves /api/status, /api/info and
/api/config/structured after the move.
