+++ src/utils/handbrake-config.js
@@ -103,40 +103,3 @@ export function validateHandBrakeConfig(config) {
     errors
   };
 }
-
-/**
- * Get default HandBrake configuration
- * @returns {Object} Default configuration object
- */
-export function getDefaultHandBrakeConfig() {
-  return {
-    enabled: false,
-    cli_path: null,
-    preset: "Fast 1080p30",
-    output_format: "mp4",
-    delete_original: false,
-    cpu_percent: 75,
-    additional_args: "",
-    subtitles: {
-      enabled: true,
-      // Include all subtitles, and prefer English by ordering it first.
-      // 'any' ensures we still pick up non-English subtitles.
-      lang_list: "eng,any",
-      all: true,
-      // Make the first selected subtitle the default (usually English when present)
-      default: "1",
-      // Keep subtitle tracks as selectable soft subtitles only.
-      burned: "none"
-    }
-  };
-}
-
-/**
- * Merge user configuration with defaults
- * @param {Object} userConfig - User provided configuration
- * @returns {Object} Merged configuration
- */
-export function mergeHandBrakeConfig(userConfig = {}) {
-  const defaults = getDefaultHandBrakeConfig();
-  return { ...defaults, ...userConfig };
-}
