+++ tests/unit/native-optical-drive.test.js
@@ -42,8 +42,6 @@ describe("NativeOpticalDrive", () => {
       expect(NativeOpticalDrive).toBeDefined();
       expect(typeof NativeOpticalDrive.ejectDrive).toBe("function");
       expect(typeof NativeOpticalDrive.loadDrive).toBe("function");
-      expect(typeof NativeOpticalDrive.ejectAllDrives).toBe("function");
-      expect(typeof NativeOpticalDrive.loadAllDrives).toBe("function");
       expect(typeof NativeOpticalDrive.isNativeAvailable).toBe("boolean"); // getter property returns boolean
     });
 
@@ -86,8 +84,7 @@ describe("NativeOpticalDrive", () => {
       mockOs.platform.mockReturnValue("win32");
     });
 
-    it("should handle native addon loading gracefully", async () => {
-      // Mock fs.existsSync to return false to simulate missing addon
+    it("reports a missing addon as unavailable rather than throwing", async () => {
       const fs = await import("fs");
       const originalExistsSync = fs.default.existsSync;
       vi.spyOn(fs.default, "existsSync").mockImplementation((path) => {
@@ -97,34 +94,21 @@ describe("NativeOpticalDrive", () => {
         return originalExistsSync(path);
       });
 
-      // When native addon can't be loaded, should throw an error
-      expect(() => {
-        const isAvailable = NativeOpticalDrive.isNativeAvailable;
-      }).toThrow("Native optical drive addon is required but failed to load");
+      expect(NativeOpticalDrive.isNativeAvailable).toBe(false);
+      // And consistently so: the first answer must match every later one.
+      expect(NativeOpticalDrive.isNativeAvailable).toBe(false);
     });
 
-    it("should handle ejectAllDrives with empty array", async () => {
-      await expect(
-        NativeOpticalDrive.ejectAllDrives([])
-      ).resolves.toBeUndefined();
-    });
+    it("gives a clean error when operating without the addon", async () => {
+      const fs = await import("fs");
+      vi.spyOn(fs.default, "existsSync").mockReturnValue(false);
 
-    it("should handle loadAllDrives with empty array", async () => {
-      await expect(
-        NativeOpticalDrive.loadAllDrives([])
-      ).resolves.toBeUndefined();
-    });
-
-    it("should continue processing drives even if some fail", async () => {
-      const drives = [{ id: "D:" }, { id: "E:" }];
-
-      // Should not throw even if individual drives fail
-      await expect(
-        NativeOpticalDrive.ejectAllDrives(drives)
-      ).resolves.toBeUndefined();
-      await expect(
-        NativeOpticalDrive.loadAllDrives(drives)
-      ).resolves.toBeUndefined();
+      await expect(NativeOpticalDrive.ejectDrive("D:")).rejects.toThrow(
+        "Native addon not available"
+      );
+      await expect(NativeOpticalDrive.loadDrive("D:")).rejects.toThrow(
+        "Native addon not available"
+      );
     });
   });
